#i122195# quicktime support is no longer available in OSX SDKs>=10.7
git-svn-id: https://svn.apache.org/repos/asf/openoffice/branches/rejuvenate01@1480387 13f79535-47bb-0310-9956-ffa450edef68

+++ main/avmedia/source/quicktime/player.hxx
@@ -99,7 +99,7 @@ private:
     int                     mnWidth;
     int                     mnHeight;
 
-    long                    mnVersion;
+    int                     mnVersion;
     oslCondition            maSizeCondition;
 };
